requirements.txt
pyserial
uptime